fix(rtp): (de)serialize buffer overflow, tested
So, I forgot that the extension, apart from the extension bytes
themselves, also contains the 4-bytes, 2 of which make up profile id,
the other 2 extension length.

Which, makes it an off-by-one error.

Also, I changed some of the manual loop-and-fill-in-the-address into
memcpy. It's nicer to look at, and faster.

And of course, rtp_header_deserialize_extension_data is done.

I will need to update the docs to add some more warnings. After all, you
can't be too careful with C.

Signed-off-by: Huy Nguyen <139140486+nguyenhuy0905@users.noreply.github.com>