fix(cmake): conan and native package interop
I made a find module for libevent. The modules conan used advises using
targets in the form `libevent::<component>`, which is technically the
modern way to do it in cmake.
cmake doesn't ship with a FindLibevent module.
No testing has beem done on devices without pkg-config and/or devices
that want to specify a custom library path.

Signed-off-by: Huy Nguyen <139140486+nguyenhuy0905@users.noreply.github.com>