feat(rtp): getters
I don't know if we want setters. Or if the client even cares about most
of these getters.
Prefereably, we want an `rtp_session` object that builds and handles the
packet header for us, and we only need to supply a few data fields (e.g.
payload type, extension, and the payload itself).

Signed-off-by: Huy Nguyen <139140486+nguyenhuy0905@users.noreply.github.com>